+++ CHANGES.rst
@@ -1,10 +1,12 @@
 Changelog
 =========
 
-1.0.1 (????-??-??)
+1.0.1 (2022-05-04)
 ------------------
 
-- ...
+- added `-s/--scale-to` option to `to-annotation-overlay-od` command to scale all annotations
+  to a predefined image size.
+- overlay output of `to-annotation-overlay-od` is now enforcing PNG format
 
 
 1.0.0 (2022-05-02)